# docs/bruno/fixtures-game-api/apispec/fixtures-game-api.openapi.yaml
...
openapi: 3.0.3
info:
  title: Fixtures Game Backend API
  version: 1.0.0
  description: API contract for the no-money football prediction game backend. Use
    `FE - ...` tags for frontend work and `Internal - ...` tags for backend jobs/admin
    maintenance.
servers:
- url: http://localhost:3001
  description: Local backend
...
paths:
...
  "/friends-groups/invite/{inviteToken}/join":
    post:
      tags:
      - FE - Groups And Invites
      summary: Join by invite link
      operationId: post_friends_groups_invite_inviteToken_join
      description: 'FE placement: invite landing page join button. Purpose: joins
        open groups immediately or creates a pending request for private groups. Notes:
        requires signed-in user.'
      x-fe-scope: frontend
      x-bruno-folder: FE App APIs - Use These In Frontend/Groups And Invites
      parameters:
      - name: inviteToken
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/JoinInviteRequest"
            example:
              message: Let me in
              userDisplayName: "{{displayName}}"
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/JoinInviteResponse"
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Forbidden for the current user.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '404':
          description: Resource not found.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '202':
          description: Private group join request created.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/JoinInviteResponse"
      security:
      - cookieAuth: []
  "/friends-groups/{friendsGroupId}/leave":
    delete:
      tags:
      - FE - Groups And Invites
      summary: Leave group
      operationId: delete_friends_groups_friendsGroupId_leave
      description: 'FE placement: group settings/member actions. Purpose: lets the
        current user leave a friends group. Members leave immediately. Owners can
        only leave when they are the final member; that archives the group and preserves
        history.'
      x-fe-scope: frontend
      x-bruno-folder: FE App APIs - Use These In Frontend/Groups And Invites
      parameters:
      - name: friendsGroupId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/LeaveGroupResponse"
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Forbidden for the current user.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '404':
          description: Resource not found.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '409':
          description: Owner must transfer ownership or remove members before leaving.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
      security:
      - cookieAuth: []
  "/friends-groups/{friendsGroupId}/members":
    get:
      tags:
      - FE - Groups And Invites
      summary: Owner lists group members
      operationId: get_friends_groups_friendsGroupId_members
      description: 'FE placement: owner member-management screen. Purpose: returns
        group members with profile previews and action flags for remove/transfer
        controls. Requires current user to be the group owner. Use matchweek overview
        for dashboard member display.'
      x-fe-scope: frontend
      x-bruno-folder: FE App APIs - Use These In Frontend/Groups And Invites
      parameters:
      - name: friendsGroupId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Owner member-management list.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/OwnerMemberListResponse"
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Current user is not the group owner.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '404':
          description: Friends group or owner membership not found.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
      security:
      - cookieAuth: []
  "/friends-groups/{friendsGroupId}/members/{userId}":
    delete:
      tags:
      - FE - Groups And Invites
      summary: Owner removes member
      operationId: delete_friends_groups_friendsGroupId_members_userId
      description: 'FE placement: owner member-management screen. Purpose: lets the
        group owner remove a member from the friends group. Owners cannot kick themselves;
        use leave instead.'
      x-fe-scope: frontend
      x-bruno-folder: FE App APIs - Use These In Frontend/Groups And Invites
      parameters:
      - name: friendsGroupId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Member removed.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/RemoveMemberResponse"
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Forbidden for the current user.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '404':
          description: Resource not found.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '409':
          description: Cannot remove group owner.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
      security:
      - cookieAuth: []
  "/friends-groups/{friendsGroupId}/transfer-ownership":
    post:
      tags:
      - FE - Groups And Invites
      summary: Transfer ownership
      operationId: post_friends_groups_friendsGroupId_transfer_ownership
      description: 'FE placement: owner member-management screen. Purpose: transfers
        group ownership to an existing member. The new owner must not already own
        another active friends group.'
      x-fe-scope: frontend
      x-bruno-folder: FE App APIs - Use These In Frontend/Groups And Invites
      parameters:
      - name: friendsGroupId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/TransferOwnershipRequest"
            example:
              newOwnerUserId: "{{newOwnerUserId}}"
      responses:
        '200':
          description: Ownership transferred.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/TransferOwnershipResponse"
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Forbidden for the current user.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '404':
          description: Resource not found.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '409':
          description: New owner already owns an active friends group.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
      security:
      - cookieAuth: []
...
  "/friends-groups/me/groups":
    get:
      tags:
      - FE - Groups And Invites
      summary: My groups
      operationId: get_friends_groups_me_groups
      description: 'FE placement: dashboard and group switcher. Purpose: lists friends
        groups current user belongs to, including invite_token and joinLink so the
        FE can rebuild share/invite UI after reload or cache clear. Also includes
        friends_group.league and friends_group.currentMatchweek so group cards can
        show labels like Premier League week 12. Includes score as the current user''s
        cumulative season/group score summary for each group.'
      x-fe-scope: frontend
      x-bruno-folder: FE App APIs - Use These In Frontend/Groups And Invites
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/MyGroupsResponse"
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Forbidden for the current user.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '404':
          description: Resource not found.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
      security:
      - cookieAuth: []
...
  "/friends-groups/invite/{inviteToken}":
    get:
      tags:
      - FE - Groups And Invites
      summary: Preview invite link
      operationId: get_friends_groups_invite_inviteToken
      description: 'FE placement: invite landing page opened from shared link. Purpose:
        validates inviteToken and displays group name, open/private access type, and
        member count before user joins.'
      x-fe-scope: frontend
      x-bruno-folder: FE App APIs - Use These In Frontend/Groups And Invites
      parameters:
      - name: inviteToken
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/InvitePreviewResponse"
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Forbidden for the current user.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '404':
          description: Resource not found.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
      security:
      - cookieAuth: []
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Supabase session cookies set by /auth/sign-in or /auth/sign-up.
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        details:
          type: object
          additionalProperties: true
          nullable: true
      required:
      - error
...
    LeaveGroupResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            status:
              type: string
              enum:
              - left
              - archived
          required:
          - status
      required:
      - data
    OwnerMember:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        role:
          type: string
          enum:
          - owner
          - member
        joinedAt:
          type: string
          format: date-time
        profile:
          "$ref": "#/components/schemas/ProfilePreview"
          nullable: true
        canRemove:
          type: boolean
          description: True when the owner can remove this member.
        canTransferOwnership:
          type: boolean
          description: True when the owner can transfer ownership to this member.
      required:
      - userId
      - role
      - joinedAt
      - profile
      - canRemove
      - canTransferOwnership
    OwnerMemberListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            "$ref": "#/components/schemas/OwnerMember"
      required:
      - data
    RemoveMemberResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            status:
              type: string
              enum:
              - removed
            friendsGroupId:
              type: string
              format: uuid
            removedUserId:
              type: string
              format: uuid
          required:
          - status
          - friendsGroupId
          - removedUserId
      required:
      - data
    TransferOwnershipRequest:
      type: object
      properties:
        newOwnerUserId:
          type: string
          format: uuid
      required:
      - newOwnerUserId
    TransferOwnershipResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            status:
              type: string
              enum:
              - transferred
            friendsGroupId:
              type: string
              format: uuid
            previousOwnerUserId:
              type: string
              format: uuid
            newOwnerUserId:
              type: string
              format: uuid
          required:
          - status
          - friendsGroupId
          - previousOwnerUserId
          - newOwnerUserId
      required:
      - data
...
    ProfilePreview:
      type: object
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
          nullable: true
        avatar_emoji:
          type: string
          nullable: true
        color_class:
          type: string
          nullable: true
      required:
      - id
...
    InvitePreview:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        accessType:
          type: string
          enum:
          - open
          - private
        memberCount:
          type: integer
      required:
      - id
      - name
      - slug
      - accessType
      - memberCount
    InvitePreviewResponse:
      type: object
      properties:
        data:
          "$ref": "#/components/schemas/InvitePreview"
      required:
      - data
    JoinInviteRequest:
      type: object
      properties:
        message:
          type: string
          nullable: true
        userDisplayName:
          type: string
          nullable: true
    JoinInviteResponse:
      type: object
      properties:
        status:
          type: string
          enum:
          - joined
          - requested
          - member
        friendsGroupId:
          type: string
          format: uuid
      required:
      - status
      - friendsGroupId
    GroupCardLeague:
      type: object
      nullable: true
      properties:
        providerLeagueId:
          type: integer
        providerSeasonId:
          type: integer
          nullable: true
        name:
          type: string
        countryName:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
        seasonName:
          type: string
          nullable: true
      required:
      - providerLeagueId
      - providerSeasonId
      - name
      - countryName
      - logoUrl
      - seasonName
    GroupCardMatchweek:
      type: object
      nullable: true
      properties:
        matchweek:
          type: string
          example: Matchweek 12
        weekNumber:
          type: integer
          nullable: true
          example: 12
        state:
          type: string
          enum:
          - upcoming
          - live
          - finished
        displayLabel:
          type: string
          example: Premier League week 12
      required:
      - matchweek
      - weekNumber
      - state
      - displayLabel
    GroupSeasonScoreSummary:
      type: object
      description: Current user's cumulative season/group score summary for this
        friends group. Values are totals across all scored matchweeks so far, not
        one selected matchweek.
      properties:
        user_id:
          type: string
          format: uuid
        fixtures_predicted:
          type: integer
        exact_score_points:
          type: integer
          description: Total exact scoreline bonus points across the season so far.
        correct_result_points:
          type: integer
          description: Total correct result points across the season so far.
        total_goals_bonus:
          type: integer
          description: Total matchweek total-goals bonus points across the season
            so far.
        red_card_bonus:
          type: integer
          description: Total red-card bonus points across the season so far.
        points_earned:
          type: integer
          description: Current user's season total points inside this friends group.
        weeks_played:
          type: integer
        rank:
          type: integer
          nullable: true
          description: Current user's season rank inside this friends group. Null
            when no matchweeks have been scored yet.
      required:
      - user_id
      - fixtures_predicted
      - exact_score_points
      - correct_result_points
      - total_goals_bonus
      - red_card_bonus
      - points_earned
      - weeks_played
      - rank
    MyGroup:
      type: object
      properties:
        role:
          type: string
          enum:
          - owner
          - member
        joined_at:
          type: string
          format: date-time
        friends_group:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            slug:
              type: string
            created_by:
              type: string
              format: uuid
            invite_token:
              type: string
              description: Stable invite token for this friends group. FE can use it
                to rebuild the invite link after reload.
            is_open:
              type: boolean
            accessType:
              type: string
              enum:
              - open
              - private
            joinLink:
              type: string
              example: "/join/abc123"
            status:
              type: string
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
            league:
              "$ref": "#/components/schemas/GroupCardLeague"
            currentMatchweek:
              "$ref": "#/components/schemas/GroupCardMatchweek"
          required:
          - id
          - name
          - slug
          - created_by
          - invite_token
          - is_open
          - accessType
          - joinLink
          - status
          - league
          - currentMatchweek
        score:
          "$ref": "#/components/schemas/GroupSeasonScoreSummary"
      required:
      - role
      - joined_at
      - friends_group
      - score
    MyGroupsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            "$ref": "#/components/schemas/MyGroup"
      required:
      - data
...